feat: add tab selector for panels, separate logic handling
