+++ src/ui/main_panel.rs
@@ -13,7 +13,6 @@ impl Default for Tab {
     }
 }
 
-#[derive(Default)]
 pub struct MainPanel {
     tab: Tab,
     send_panel: SendPanel,
@@ -21,14 +20,33 @@ pub struct MainPanel {
 }
 
 impl MainPanel {
+    pub fn new() -> Self {
+        Self {
+            tab: Tab::default(),
+            send_panel: SendPanel::default(),
+            receive_panel: ReceivePanel::default(),
+        }
+    }
+}
+
+impl MainPanel {
     pub fn ui(&mut self, ctx: &egui::Context) {
         egui::TopBottomPanel::top("tabs").show(ctx, |ui| {
             ui.horizontal(|ui| {
-                if ui.selectable_label(self.tab == Tab::Send, "Send").clicked() {
+                if ui
+                    .selectable_label(
+                        self.tab == Tab::Send,
+                        egui::RichText::new("Send").size(20.0).strong(),
+                    )
+                    .clicked()
+                {
                     self.tab = Tab::Send;
                 }
                 if ui
-                    .selectable_label(self.tab == Tab::Receive, "Receive")
+                    .selectable_label(
+                        self.tab == Tab::Receive,
+                        egui::RichText::new("Receive").size(20.0).strong(),
+                    )
                     .clicked()
                 {
                     self.tab = Tab::Receive;
@@ -38,7 +56,7 @@ impl MainPanel {
 
         egui::CentralPanel::default().show(ctx, |ui| match self.tab {
             Tab::Send => self.send_panel.ui(ui, ctx),
-            Tab::Receive => self.receive_panel.ui(ui),
+            Tab::Receive => self.receive_panel.ui(ui, ctx),
         });
     }
 }
